requirements.txt
asttokens==3.0.1
bidict==0.23.1
build==1.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.11
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
pip-tools==7.5.2
polars==1.35.2
polars-runtime-32==1.35.2
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-socketio==5.14.3
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
supervision==0.27.0
sympy==1.14.0
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.228
ultralytics-thop==2.0.18
urllib3==2.5.0
wcwidth==0.2.14
websocket-client==1.9.0
wsproto==1.3.1
